debug

+++ templates/admin/shakhes.ruydad.php
@@ -11,7 +11,7 @@
                 echo $msg;
             }
             ?>
-            <?php if ($this->time['import_time'] == 1): ?>
+            
             <div class="alert alert-warning">زمان اتمام: <?=convertDate($this->time['finish_date'])?></div>
             <form action="<?= RELA_DIR ?>admin/?component=shakhes&action=ruydad" method="post">
                 <table class="form">
@@ -141,9 +141,7 @@
                 <button name="temporary" value="1" class="btn btn-warning btn-large">ثبت موقت</button>
             
             </form>
-            <? else:?>
-                                <div class="alert alert-warning">زمان ارسال اطلاعات به پایان رسیده است</div>
-                                <? endif;?>
+            
         </div>
         <div class="panel-heading bg-green">
             <h3 class="panel-title rtl "> لیست رویداد</h3>
@@ -206,18 +204,20 @@
                             <td><?= readMore($v['tozihat'],10) ?></td>
 
                             <td>
-                                <? if( in_array($admin_info['admin_id'], [$v['import_admin'],$v['admin_id']]) ):?>
-                                    <? if(($v['status'] == 0 || $v['status'] == 1) ):  ?>
-                                    <form action="<?= RELA_DIR ?>admin/?component=shakhes&action=ruydad" method="post">
-                                        <button name="sendToParent" value="<?= $v['id'] ?>" onclick="confirm('آیا از ارسال به مافوق مطمئن هستید؟')"
-                                                class="btn btn-xs btn-block btn-success pull-right">ارسال به مافوق</button>
-                                    </form>
-                                    <a href="<?= RELA_DIR ?>admin/?component=shakhes&action=ruydad&method=delete&id=<?= $v['id'] ?>"
-                                       class="btn btn-xs btn-block btn-danger pull-right" onclick="return confirm('آیا برای حذف مطمئن هستید؟')">حذف</a>
-                                    <? else:?>
-                                        <?= ($v['status'] == 2) ? 'ارسال به مافوق' : '' ?>
-                                        <?= ($v['status'] == 3) ? 'تایید توسط مافوق' : '' ?>
-                                        <?= ($v['status'] == 4) ? 'تایید نهایی ' : '' ?>
+                                <?php if ($this->time['import_time'] == 1): ?>
+                                    <? if( in_array($admin_info['admin_id'], [$v['import_admin'],$v['admin_id']]) ):?>
+                                        <? if(($v['status'] == 0 || $v['status'] == 1) ):  ?>
+                                        <form action="<?= RELA_DIR ?>admin/?component=shakhes&action=ruydad" method="post">
+                                            <button name="sendToParent" value="<?= $v['id'] ?>" onclick="confirm('آیا از ارسال به مافوق مطمئن هستید؟')"
+                                                    class="btn btn-xs btn-block btn-success pull-right">ارسال به مافوق</button>
+                                        </form>
+                                        <a href="<?= RELA_DIR ?>admin/?component=shakhes&action=ruydad&method=delete&id=<?= $v['id'] ?>"
+                                        class="btn btn-xs btn-block btn-danger pull-right" onclick="return confirm('آیا برای حذف مطمئن هستید؟')">حذف</a>
+                                        <? else:?>
+                                            <?= ($v['status'] == 2) ? 'ارسال به مافوق' : '' ?>
+                                            <?= ($v['status'] == 3) ? 'تایید توسط مافوق' : '' ?>
+                                            <?= ($v['status'] == 4) ? 'تایید نهایی ' : '' ?>
+                                        <? endif;?>
                                     <? endif;?>
                                 <? endif;?>
